refactor: extract study swipe workflow
