+++ src/features/study/hooks/useStudy.ts
@@ -1,13 +1,8 @@
-import { useAuthUid } from "@/entities/auth";
 import type { Card } from "@/entities/card";
 import type { DeckId } from "@/entities/deck";
 import { type SwipeDirection, usePreferences } from "@/entities/preferences";
-import { editStudyProgress } from "@/entities/study-progress";
 import {
   calculateStudySessionIndex,
-  getStudySession,
-  moveStudySession,
-  planStudySessionSwipe,
   removeStudySession,
   resolveStudySession,
   setStudySessionIndex,
@@ -18,7 +13,7 @@ import {
 
 import * as React from "react";
 
-import { useSwipeFeedback } from "./useSwipeFeedback";
+import { useSwipe } from "../model/useSwipe";
 
 interface StudyCommands {
   swipeUp: () => Promise<void>;
@@ -47,43 +42,13 @@ export type StudyState = StudyCommands &
   );
 
 export const useStudy = (deckId: DeckId, cards: readonly Card[], onInvalid: () => void): StudyState => {
-  const uid = useAuthUid();
   const preferences = usePreferences();
   const [showBackText, setShowBackText] = React.useState(false);
   const [autoPlay, setAutoPlay] = React.useState(preferences.study.defaultAutoPlay);
   const hideBackText = () => setShowBackText(false);
   const toggleBackText = () => setShowBackText((visible) => !visible);
   const toggleAutoPlay = () => setAutoPlay((playing) => !playing);
-  const feedback = useSwipeFeedback(preferences.appearance.showSwipeFeedback);
-  const saveProgress = (progress: Parameters<typeof editStudyProgress>[1]) => editStudyProgress(uid, progress);
-  const swipeState = React.useRef<{ inProgress: boolean }>({ inProgress: false });
-  const swipe = async (direction: SwipeDirection): Promise<void> => {
-    // biome-ignore lint/suspicious/noUnnecessaryConditions: The awaited write lets another event enter this closure.
-    if (swipeState.current.inProgress) return;
-
-    const swipeAction = preferences.controls[direction];
-    const swipePlan = planStudySessionSwipe(getStudySession(deckId), cards, swipeAction, Date.now());
-    if (swipePlan.effect === "none") return;
-    if (swipePlan.effect === "exit") {
-      feedback.showSwipe(direction);
-      removeStudySession(deckId);
-      return;
-    }
-
-    swipeState.current.inProgress = true;
-    // The visible card advances only after persistence succeeds, so failed writes need no session rollback.
-    const saved = await saveProgress(swipePlan.progress).then(
-      () => true,
-      () => false
-    );
-    swipeState.current.inProgress = false;
-    if (!saved) return;
-
-    if (!moveStudySession(swipePlan.session, swipePlan.effect)) return;
-
-    feedback.showSwipe(direction);
-    if (preferences.appearance.hideBodyWhenCardChanged) hideBackText();
-  };
+  const swipe = useSwipe(deckId, cards, hideBackText);
   const updateIndex = (currentIndex: number): void => {
     if (!setStudySessionIndex(deckId, currentIndex)) return;
     hideBackText();
@@ -126,10 +91,10 @@ export const useStudy = (deckId: DeckId, cards: readonly Card[], onInvalid: () =
     return () => window.clearTimeout(timeout);
   }, [autoPlay, deckId, preferences.study.cardInterval, resolvedSession.status, session]);
   const commands: StudyCommands = {
-    swipeUp: () => swipe("cardSwipeUp"),
-    swipeDown: () => swipe("cardSwipeDown"),
-    swipeLeft: () => swipe("cardSwipeLeft"),
-    swipeRight: () => swipe("cardSwipeRight"),
+    swipeUp: swipe.swipeUp,
+    swipeDown: swipe.swipeDown,
+    swipeLeft: swipe.swipeLeft,
+    swipeRight: swipe.swipeRight,
     toggleBackText,
     toggleAutoPlay,
   };
@@ -147,6 +112,6 @@ export const useStudy = (deckId: DeckId, cards: readonly Card[], onInvalid: () =
     showSwipeButtonList: preferences.controls.showSwipeButtonList,
     autoPlay,
     updateIndex,
-    ...(feedback.lastSwipe !== undefined ? { swipeFeedback: feedback.lastSwipe } : {}),
+    ...(swipe.swipeFeedback !== undefined ? { swipeFeedback: swipe.swipeFeedback } : {}),
   };
 };
